# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Task API
  description: >
    A small, multi-user HTTP REST API for task management, built with Go
    using only the standard library plus pgx (PostgreSQL) and
    golang.org/x/crypto (bcrypt). Every task belongs to exactly one user
    and is only ever visible to that user. Each task has a title, an
    optional description, a status (`pending`, `in_progress`, `done`, or
    `cancelled`, following the transition rules below), and a priority
    (`low`, `medium`, or `high`). Authentication is opaque bearer session
    tokens, issued by POST /auth/login.


    This document describes exactly the behavior implemented in
    `internal/task/handler.go`+`service.go`+`repository.go` and
    `internal/user/handler.go`+`service.go`+`repository.go` as of the
    current codebase — no aspirational or planned behavior is included.


    Every response from every endpoint — including error responses and
    responses produced before routing, such as a `404` for an unknown path
    — additionally carries a fixed set of security headers
    (`middleware.SecurityHeaders`): `Content-Security-Policy: default-src
    'none'; frame-ancestors 'none'`, `X-Frame-Options: DENY`, and
    `X-Content-Type-Options: nosniff`. They are constant, identical
    everywhere, and are not part of the API contract a client programs
    against, so they are documented here once rather than repeated on each
    response. `Strict-Transport-Security` joins them only when the
...
servers:
...
security:
  - bearerAuth: []
paths:
...
  /tasks:
    post:
      tags: [Tasks]
      operationId: createTask
      summary: Create a task
      description: >
        Creates a new task owned by the authenticated caller, with status
        `pending`. `id`, `status`, `created_at` and `updated_at` are
        always assigned by the server; if the client includes any of
        these fields in the request body, or any other field not part of
        CreateTaskRequest, they are silently ignored — the API never
        returns a 400 for unknown fields.
      requestBody:
        required: true
        description: >
          Request bodies larger than 1 MiB are rejected (surfaces as the
          same 400 "invalid request body" response as malformed JSON).
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTaskRequest'
            examples:
              valid:
                summary: Valid request
                value:
                  title: Buy groceries
                  description: Milk, eggs, bread
                  priority: high
              titleOnly:
                summary: Description and priority omitted (both optional)
                value:
                  title: Buy groceries
      responses:
        '201':
          description: Task created successfully.
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
              example:
                id: f47ac10b-58cc-4372-a567-0e02b2c3d479
                title: Buy groceries
                description: Milk, eggs, bread
                status: pending
                priority: high
                created_at: "2026-08-08T12:00:00Z"
                updated_at: "2026-08-08T12:00:00Z"
        '400':
          description: >
            Malformed JSON body; a missing, empty, or whitespace-only
            `title`; a `title`/`description` longer than 200/2000
            characters (after trimming); or a `priority` that isn't one
            of `low`/`medium`/`high`.
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                malformedJSON:
                  summary: Malformed JSON
                  value:
                    error: invalid request body
                emptyTitle:
                  summary: Empty or whitespace-only title
                  value:
                    error: "invalid input: title must not be empty"
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          description: >
            A task with the server-generated id already exists.
            Practically unreachable in normal operation: `id` is a
            UUIDv4 generated internally by the Service and never
            supplied by the client, so this can only happen on an
            astronomically unlikely UUID collision. Documented because
            the code path exists and is reachable in principle, not
            because it is expected to occur.
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: task already exists
        '500':
          $ref: '#/components/responses/InternalServerError'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'

    get:
      tags: [Tasks]
      operationId: listTasks
      summary: List the authenticated caller's tasks
      description: >
        Returns the authenticated caller's own tasks — never another
        user's — ordered by `created_at` ascending (oldest first; ties
        broken by `id`). The Repository guarantees this ordering and
        applies the ownership filter and `limit`/`offset` itself (pushed
        into the SQL query for the PostgreSQL-backed store: `WHERE
        user_id = … ORDER BY … LIMIT … OFFSET …`), rather than the
        Service fetching everything and filtering/paginating in memory.
        `limit` and `offset` are optional; omitting both returns every
        task the caller owns. Returns an empty JSON array, never `null`,
        when the caller has no tasks or when `offset` is past the end of
        the list.
      parameters:
        - name: limit
          in: query
          required: false
          description: >
            Maximum number of tasks to return. Sending the parameter with
            an empty value (`?limit=`) is treated as omitting it, not as
            an error.
          schema:
            type: integer
            minimum: 0
          example: 20
        - name: offset
          in: query
          required: false
          description: >
            Number of tasks to skip before collecting `limit` results.
            Sending the parameter with an empty value (`?offset=`) is
            treated as omitting it, not as an error.
          schema:
            type: integer
            minimum: 0
          example: 0
      responses:
        '200':
          description: >
            Success. The array is empty (`[]`) if the caller has no
            tasks, or if `offset` is past the end of the list.
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestID'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Task'
              examples:
                withData:
                  summary: Tasks exist
                  value:
                    - id: f47ac10b-58cc-4372-a567-0e02b2c3d479
                      title: Buy groceries
                      description: Milk, eggs, bread
                      status: pending
                      priority: high
                      created_at: "2026-08-08T12:00:00Z"
                      updated_at: "2026-08-08T12:00:00Z"
                empty:
                  summary: No tasks exist
                  value: []
        '400':
          description: '`limit` or `offset` is present but not a non-negative integer.'
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: "invalid input: limit must be a non-negative integer"
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /tasks/{id}:
    get:
      tags: [Tasks]
      operationId: getTask
      summary: Get one of the authenticated caller's tasks by id
      parameters:
        - $ref: '#/components/parameters/TaskID'
      responses:
        '200':
          description: Task found.
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
              example:
                id: f47ac10b-58cc-4372-a567-0e02b2c3d479
                title: Buy groceries
                description: Milk, eggs, bread
                status: pending
                priority: high
                created_at: "2026-08-08T12:00:00Z"
                updated_at: "2026-08-08T12:00:00Z"
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'

    put:
      tags: [Tasks]
      operationId: updateTask
      summary: Update a task's title, description and/or priority
      description: >
        Replaces `title` and `description`, and — if `priority` is
        present and non-empty — `priority` too; an omitted or empty
        `priority` leaves it unchanged. `id`, `status` and `created_at`
        are never modified by this endpoint — to change `status`, use
        PATCH /tasks/{id}/status (or PATCH /tasks/{id}/done). As with
        creation, any field in the body other than
        `title`/`description`/`priority` (e.g. a client-supplied `status`
        or `id`) is silently ignored. `updated_at` is always refreshed to
        the current time.
      parameters:
        - $ref: '#/components/parameters/TaskID'
      requestBody:
        required: true
        description: >
          Request bodies larger than 1 MiB are rejected (surfaces as the
          same 400 "invalid request body" response as malformed JSON).
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTaskRequest'
            example:
              title: Buy groceries and cook dinner
              description: Don't forget olive oil
      responses:
        '200':
          description: Task updated.
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
              example:
                id: f47ac10b-58cc-4372-a567-0e02b2c3d479
                title: Buy groceries and cook dinner
                description: Don't forget olive oil
                status: pending
                priority: high
                created_at: "2026-08-08T12:00:00Z"
                updated_at: "2026-08-08T12:05:00Z"
        '400':
          description: >
            Malformed JSON body; a missing, empty, or whitespace-only
            `title`; a `title`/`description` longer than 200/2000
            characters (after trimming); or a non-empty `priority` that
            isn't one of `low`/`medium`/`high`.
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                malformedJSON:
                  summary: Malformed JSON
                  value:
                    error: invalid request body
                emptyTitle:
                  summary: Empty or whitespace-only title
                  value:
                    error: "invalid input: title must not be empty"
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: >
            Optimistic concurrency conflict: another request updated this
            task between when this request's caller last read it and this
            write. The caller should re-fetch the task and retry.
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: task was modified concurrently, please retry
        '500':
          $ref: '#/components/responses/InternalServerError'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'

    delete:
      tags: [Tasks]
      operationId: deleteTask
      summary: Delete a task
      description: Permanently removes a task. No request body.
      parameters:
        - $ref: '#/components/parameters/TaskID'
      responses:
        '204':
          description: Task deleted successfully. No response body.
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestID'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /tasks/{id}/status:
    patch:
      tags: [Tasks]
      operationId: transitionTaskStatus
      summary: Move a task to a different status
      description: >
        Moves the task to `status`, if that's a legal transition from its
        current status — see the Status schema's transition table.
        Requesting the task's *current* status again always succeeds as a
        no-op (200, `updated_at` unchanged, nothing written) — the same
        idempotency PATCH /tasks/{id}/done has always provided for an
        already-`done` task, generalized to every status.
      parameters:
        - $ref: '#/components/parameters/TaskID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransitionStatusRequest'
            example:
              status: in_progress
      responses:
        '200':
          description: Task moved to the new status (or already was — same response shape either way).
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
              example:
                id: f47ac10b-58cc-4372-a567-0e02b2c3d479
                title: Buy groceries
                description: Milk, eggs, bread
                status: in_progress
                priority: high
                created_at: "2026-08-08T12:00:00Z"
                updated_at: "2026-08-08T12:05:00Z"
        '400':
          description: Malformed JSON body, or `status` is not one of the four known values.
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: 'invalid input: unknown status "archived"'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: >
            Either an optimistic concurrency conflict (see PUT
            /tasks/{id}'s 409), or the requested transition is not legal
            from the task's current status (e.g. `cancelled` →
            `done`) — distinguishable by the error message, both map to
            409 since both mean "this request conflicts with the
            resource's current state."
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                invalidTransition:
                  summary: Illegal transition
                  value:
                    error: 'invalid status transition: cannot move from "cancelled" to "done"'
                concurrencyConflict:
                  summary: Optimistic concurrency conflict
                  value:
                    error: task was modified concurrently, please retry
        '500':
          $ref: '#/components/responses/InternalServerError'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /tasks/{id}/done:
    patch:
      tags: [Tasks]
      operationId: completeTask
      summary: Mark a task as done
      description: >
        Marks the task `done`. No request body is read by this endpoint —
        any body sent is ignored. A thin, backward-compatible convenience
        wrapper around PATCH /tasks/{id}/status with `{"status":"done"}`
        — kept as its own endpoint because it predates the general one and
        its wire contract must not change.


        This operation is **idempotent**: calling it on a task that is
        already `done` returns 200 with the task unchanged, and does
        *not* refresh `updated_at` or write to the repository.

        It is *not* a universal shortcut: it applies the same transition
        rules as PATCH /tasks/{id}/status, so it only succeeds from a
        status that can legally reach `done` — `pending` and
        `in_progress`. A `cancelled` task cannot be completed in one hop
        (see the Status schema's transition table) and returns 409; move
        it back to `pending` first.
      parameters:
        - $ref: '#/components/parameters/TaskID'
      responses:
        '200':
          description: >
            Task marked as done, or already was — response shape is the
            same either way.
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
              example:
                id: f47ac10b-58cc-4372-a567-0e02b2c3d479
                title: Buy groceries
                description: Milk, eggs, bread
                status: done
                priority: high
                created_at: "2026-08-08T12:00:00Z"
                updated_at: "2026-08-08T12:05:00Z"
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: >
            Either the task's current status cannot legally reach `done`
            in one hop — only `cancelled` can't (see the Status schema's
            transition table); move it back to `pending` first — or an
            optimistic concurrency conflict: another request updated this
            task between when this request's caller last read it and this
            write, in which case the caller should re-fetch and retry.

            Neither is ever returned when the task is already `done`:
            that path is idempotent and never writes.
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                invalidTransition:
                  summary: Task is cancelled, which cannot reach done directly
                  value:
                    error: 'complete task: invalid status transition: cannot move from "cancelled" to "done"'
                concurrencyConflict:
                  summary: Optimistic concurrency conflict
                  value:
                    error: task was modified concurrently, please retry
        '500':
          $ref: '#/components/responses/InternalServerError'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: >
        Opaque session token returned by POST /auth/login (not a JWT — it
        has no independently verifiable structure; the server looks it up
        by its hash on every request). Send as `Authorization: Bearer
        <token>`.
...
  headers:
    XRequestID:
      description: >
        Correlates this response with server-side structured logs
        (`request_id` field). Echoes the client-supplied `X-Request-Id`
        request header when it is present and well-formed (non-empty,
        at most 128 characters, and restricted to `[A-Za-z0-9_-]`);
        otherwise the server generates a new one. Present on every
        response from every endpoint, including error responses.
      schema:
        type: string
      example: 7a947bb6436a2e5009efcaa55c42cdf8
  schemas:
...
    Task:
      type: object
      description: >
        The persisted representation of a task, as returned by every
        endpoint that returns a task body. Field order and JSON keys
        match the `json` struct tags on task.Task exactly (snake_case).
        The owning user's id is deliberately not included — every task a
        client can ever retrieve already belongs to it.
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
          description: Server-generated UUIDv4. Never accepted from the client.
          example: f47ac10b-58cc-4372-a567-0e02b2c3d479
        title:
          type: string
          example: Buy groceries
        description:
          type: string
          example: Milk, eggs, bread
        status:
          $ref: '#/components/schemas/Status'
        priority:
          $ref: '#/components/schemas/Priority'
        created_at:
          type: string
          format: date-time
          readOnly: true
          description: Set once at creation; never modified afterwards.
          example: "2026-08-08T12:00:00Z"
        updated_at:
          type: string
          format: date-time
          readOnly: true
          description: >
            Refreshed on every mutation that actually changes something
            (PUT, and every status transition that isn't a same-status
            no-op). Left unchanged by an idempotent PATCH .../done or
            PATCH .../status call.
          example: "2026-08-08T12:00:00Z"
      required:
        - id
        - title
        - description
        - status
        - priority
        - created_at
        - updated_at
    CreateTaskRequest:
      type: object
      description: >
        Accepted body for POST /tasks. Any additional field beyond
        `title`/`description`/`priority` is silently ignored by the
        server rather than rejected. `title`/`description` are trimmed of
        leading/trailing whitespace before being stored (length limits
        below are checked after trimming).
      properties:
        title:
          type: string
          maxLength: 200
          description: >
            Required. Rejected with 400 if empty or made of only
            whitespace after trimming, or if longer than 200 characters.
          example: Buy groceries
        description:
          type: string
          maxLength: 2000
          description: >
            Optional. Defaults to an empty string if omitted. Rejected
            with 400 if longer than 2000 characters.
          example: Milk, eggs, bread
        priority:
          type: string
          enum: [low, medium, high, ""]
          description: >
            Optional. Defaults to `medium` when omitted **or** sent as an
            empty string — the server treats "" as "not provided" rather
            than rejecting it, which is why "" appears in this enum even
            though it is not a valid stored Priority value.
          example: high
      required:
        - title
...
    TransitionStatusRequest:
      type: object
      description: Accepted body for PATCH /tasks/{id}/status.
      properties:
        status:
          $ref: '#/components/schemas/Status'
      required:
        - status
    ErrorResponse:
      type: object
      description: >
        The single JSON envelope used by every error response from every
        endpoint in this API.
      properties:
        error:
          type: string
          description: Human-readable error message. Never contains internal details.
          example: task not found
      required:
        - error
  responses:
    Unauthorized:
      description: >
        Missing, malformed, or invalid/expired `Authorization: Bearer`
        header.
      headers:
        X-Request-Id:
          $ref: '#/components/headers/XRequestID'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: invalid or expired session token
...